Organize components into folder
The project has grown enough that a /components dir is helpful

Refs #40
